+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.20.0 - in development
 - [Command to Toggle View Annotations #70](https://github.com/pedrorgirardi/Pep/issues/70)
+- Removed Command to Clear a View's Annotations
 
 ## 0.19.0 - 2023-10-31
 There were many internal changes in this version and I'm sorry I made some breaking changes.